requirements.txt
crewai
crewai_tools
openai
streamlit
langsmith
langchain
langchain-community
chromadb
PyPDF2
python-dotenv
openai
pypdf
tiktoken
langchain-openai 
langchain-chroma
fitz
PyMuPDF